Pair breeders now works - but with a hack
Made add_source work, had to hack it to allow empty entry into source table
I think Background should be a secondary key in the Strain table. Otherwise we need to include a menu for it in PairBreeders (or look it up from them I guess), but I don't think there should be multiple Strains with the same name but different Backgrounds.